refactor: Move publish plugin to own module
