+++ infrastructure-base/src/main/kotlin/dsl/Predicates.kt
@@ -1,0 +1,5 @@
+package com.redmadrobot.build.dsl
+
+/** Returns `true` if build is running on CI. */
+public val isRunningOnCi: Boolean
+    get() = System.getenv("CI") == "true"
