+++ infrastructure-publish/src/main/kotlin/dsl/PublishingPredicates.kt
@@ -2,10 +2,6 @@ package com.redmadrobot.build.dsl
 
 import org.gradle.api.Project
 
-/** Returns `true` if build is running on CI. */
-public val isRunningOnCi: Boolean
-    get() = System.getenv("CI") == "true"
-
 /**
  * Checks that credentials for the repository with given [name] are exist.
  *
