+++ infrastructure-publish/src/main/kotlin/package.kt
@@ -1,0 +1,1 @@
+package com.redmadrobot.build
